+++ scripts/jsxbin.js
@@ -1,4 +1,4 @@
-import { renameSync } from 'fs'
+import { renameSync, existsSync } from 'fs'
 import { readdir } from 'fs/promises'
 import { fork } from 'child_process'
 import readdirp from 'readdirp'
@@ -20,8 +20,18 @@ scripts.forEach((script) => {
 
 async function getExtensionPath() {
     const extensionsPath = path.join(homedir(), '.vscode', 'extensions')
+    if (!existsSync(extensionsPath)) {
+        throw new Error(`Missing VSCode extensions folder at ${extensionsPath}`)
+    }
     const extensions = await readdir(extensionsPath)
     const extensionName = 'adobe.extendscript-debug'
-    const extendscript = extensions.find((f) => f.includes(extensionName))
-    return path.join(extensionsPath, extendscript, 'public-scripts', 'exportToJSXBin.js')
+    const extendscriptFolder = extensions.find((f) => f.includes(extensionName))
+    if (!extendscriptFolder) {
+        throw new Error(`Missing VSCode extension ${path.join(extensionsPath, extensionName)}`)
+    }
+    const jsxBinPath = path.join(extensionsPath, extendscriptFolder, 'public-scripts', 'exportToJSXBin.js')
+    if (!existsSync(jsxBinPath)) {
+        throw new Error(`Expected script at ${jsxBinPath}`)
+    }
+    return jsxBinPath
 }